+++ .gitignore
@@ -1,2 +1,2 @@
-FirstFantasy/src/__pycache__/*
-FirstFantasy/src/saves/*
+src/__pycache__/*
+src/saves/*
