+++ README.txt
@@ -1,0 +1,22 @@
+The First Fantasy is a text-based role-playing game developed completely with Python 2.7.3 for use in a console window. In the game, you can travel around the world, fighting enemies to gain experience, gold, and loot to improve your character. A clear goal is yet to be defined, but the challenge of progressing farther away from your starting area provides incentive to play.
+
+To Run:
+
+Windows Only:
+Execute TheFirstFantasy.exe within the dist directory.
+
+Any OS with Python 2.7.3:
+Change directory into the src folder and type the command: python main.py 
+
+To Build Distributable on Windows:
+
+Python and py2exe is needed create a new dist. Once installed, run build.bat.
+
+Features:
+
+-Randomized world creation ensuring that each new game provides a different experience.
+-Randomized enemies and difficulty.
+-Randomly created weapons and armor to equip.
+-Simple interface.
+-Save and load progress.
+-No compilation necessary. All ready-to-go out of the box.
