+++ src/characters.py
